Move lib dir inside src dir
