+++ src/lib/helperFunctions.js
@@ -1,9 +1,9 @@
 import axios from "axios";
-import { aboutMe, itemsToFetch, includedRepos } from "../src/constants";
+import { aboutMe, itemsToFetch, includedRepos } from "../constants";
 
 export const scrollToSection = (id) => {
   const element = document.getElementById(id);
-  const yOffset = -70; // Adjust this value based on your header height
+  const yOffset = -70;
   const y = element.getBoundingClientRect().top + window.scrollY + yOffset;
 
   window.scrollTo({ top: y, behavior: "smooth" });
